% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEastNorth.R
\name{calcEastNorth}
\alias{calcEastNorth}
\title{Convert decimalLatitude and decimalLongitude to utms. Multiple UTM zones are allowed.}
\usage{
calcEastNorth(latitude, longitude, utmZone)
}
\arguments{
\item{latitude}{Decimal latitude [numeric]}

\item{longitude}{Decimal longitude [numeric]}

\item{utmZone}{utmZones corresponding to the coordinates [character]}
}
\value{
A data frame with 2 columns: easting, northing
}
\description{
Conversion Function. Given vectors of decimalLatitude, decimalLongitude, and UTM zone, return corresponding easting and northing.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
