% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEastNorthByZone.R
\name{calcEastNorthByZone}
\alias{calcEastNorthByZone}
\title{Convert decimalLatitude and decimalLongitude to utms, one zone at a time.}
\usage{
calcEastNorthByZone(latitude, longitude, utmZone)
}
\arguments{
\item{latitude}{Decimal latitude [numeric]}

\item{longitude}{Decimal longitude [numeric]}

\item{utmZone}{A single utmZone in form "17N" [character]}
}
\value{
A data frame with 2 columns: easting, northing
}
\description{
Conversion Function. Given vectors of decimalLatitude and decimalLongitude., and a UTM zone, return corresponding easting and northing.
}
\examples{
calcEastNorthByZone(latitude=45.8, longitude=-101, '17N')
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
