% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLatLong.R
\name{calcLatLong}
\alias{calcLatLong}
\title{Convert utms to decimalLatitude and decimalLongitude. Multiple UTM zones are allowed.}
\usage{
calcLatLong(easting, northing, utmZone)
}
\arguments{
\item{easting}{Easting UTM coordinates [numeric]}

\item{northing}{Northing UTM coordinates [numeric]}

\item{utmZone}{UTM zones corresponding to the coordinates [character]}
}
\value{
A data frame with 2 columns: decimalLatitude, decimalLongitude
}
\description{
Conversion Function. Given vectors of easting, northing, and UTM zone, return corresponding decimalLatitude and decimalLongitude. This conversion function wraps calcLatLongByZone to apply over multiple utmZones.
}
\examples{
calcLatLong(easting=c(747725.21, 735000), 
     northing=c(4308017.75, 4307000), 
     utmZone=c('17N','17N'))
calcLatLong(easting=c(747725.21, 581417.95, NA), 
     northing=c(4308017.75, 5074637.09, NA), 
     utmZone=c('17N','10N','10N'))
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
