% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLatLongByZone.R
\name{calcLatLongByZone}
\alias{calcLatLongByZone}
\title{Convert utms to decimalLatitude and decimalLongitude, one zone at a time.}
\usage{
calcLatLongByZone(easting, northing, utmZone)
}
\arguments{
\item{easting}{Easting UTM coordinates [numeric]}

\item{northing}{Northing UTM coordinates [numeric]}

\item{utmZone}{A single utmZone in form "17N" [character]}
}
\value{
A data frame with 2 columns: decimalLatitude, decimalLongitude
}
\description{
Conversion Function. Given vectors of easting and northing, and a UTM zone, return corresponding decimalLatitude and decimalLongitude.
}
\examples{
calcLatLongByZone(easting=747725.21, northing=4308017.75, '17N')
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
