% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipCell.doc.R
\docType{data}
\name{clipCell}
\alias{clipCell}
\title{Documentation for clipCell data set}
\format{
A data frame with 1080 rows and 4 variables
\describe{
  \item{pointID}{identifier of point within plot}
  \item{clipCellNumber}{identifier of clip cell within point}
  \item{offsetEasting}{offset to be added to easting value of southwest corner of plot}
  \item{offsetNorthing}{offset to be added to northing value of southwest corner of plot}
  }
}
\usage{
clipCell
}
\description{
Location offsets within NEON plots for clip cells used to collect herbaceous biomass, litterfall, and bryophytes.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Courtney Meier \email{cmeier@battelleecology.org}
}
\keyword{datasets}
