% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBLAN.R
\name{convertBLAN}
\alias{convertBLAN}
\title{Convert easting and northing coordinates to align BLAN and LEWI locations with AOP flights}
\usage{
convertBLAN(data, easting = "easting", northing = "northing")
}
\arguments{
\item{data}{A data frame including columns of easting and northing coordinates [data frame]}

\item{easting}{Column name containing easting values [character]}

\item{northing}{Column name containing northing values [character]}
}
\value{
The input data frame, with all easting and northing coordinates in 17N
}
\description{
Converts easting and northing coordinates from 18N to 17N to align BLAN and LEWI locations with AOP flights.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
