% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.calc.geo.os.R
\name{def.calc.geo.os}
\alias{def.calc.geo.os}
\title{Calculate more precise geolocations for specific NEON data products}
\usage{
def.calc.geo.os(data, dataProd)
}
\arguments{
\item{data}{A data frame containing either NEON named locations or geolocations. Field names of locations must match standard NEON location field names.}

\item{dataProd}{The table name of the NEON data product table to find locations for. Must be one of: ltr_pertrap, sls_soilCoreCollection, brd_perpoint (list will continue to expand over time)}
}
\value{
A data frame of geolocations for the input product and data
}
\description{
Calculation Function. Refine the geolocation data associated with NEON data products, based on product-specific rules and spatial designs.
}
\examples{
d <- data.frame(namedLocation="GUAN_044.basePlot.ltr", subplotID=23, trapID="GUAN_044_385")
def.calc.geo.os(d, "ltr_pertrap")
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
