% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.calc.latlong.by.zone.R
\name{def.calc.latlong.by.zone}
\alias{def.calc.latlong.by.zone}
\title{Convert utms to decimalLatitude and decimalLongitude, one zone at a time.}
\usage{
def.calc.latlong.by.zone(df, utmZone)
}
\arguments{
\item{\code{df}}{A data frame columns labeled easting, northing.}

\item{\code{utmZone}}{The utmZone of all rows of the dataframe to be converted in form "17N"}
}
\value{
The original data frame with decimalLatitude and decimalLongitude calculated.
}
\description{
Conversion Function. Given a data frame with geolocations in UTMs, return corresponding decimalLatitude and decimalLongitude. Original decimalLatitude and decimalLongitude will be removed. This conversion function works only with a datafram with a single (unique) utmZone. Use def.utm.tolatlong to apply over variable utmZones.
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Sarah Elmendorf \email{selmendorf@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
