% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.calc.latlong.R
\name{def.calc.latlong}
\alias{def.calc.latlong}
\title{Convert utms to decimalLatitude and decimalLongitude, one zone at a time.}
\usage{
def.calc.latlong(df)
}
\arguments{
\item{df}{A data frame columns labeled easting, northing.}
}
\value{
The original data frame with decimalLatitude and decimalLongitude calculated.
}
\description{
Conversion Function. Given a data frame with geolocations in UTMs, return corresponding decimalLatitude and decimalLongitude. Original decimalLatitude and decimalLongitude will be removed. This conversion function wraps def.calc.utm.tolatlong.by.zone to apply over variable utmZones.
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Sarah Elmendorf \email{selmendorf@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
