% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.extr.geo.os.R
\name{def.extr.geo.os}
\alias{def.extr.geo.os}
\title{Extract geolocation data from NEON API}
\usage{
def.extr.geo.os(data, locCol = "namedLocation")
}
\arguments{
\item{data}{A data frame in which one column contains the named locations}

\item{locCol}{The column name of the column containing the named locations. Defaults to namedLocation}
}
\value{
A data frame of the geolocation data for the input named locations
}
\description{
Definition Function. Extract geolocation data from the NEON API, for a given set of named location values
}
\examples{
d <- data.frame(namedLocation=c("GUAN_044.basePlot.ltr","GRSM_003.birdGrid.brd"), otherData=c(1,2))
def.extr.geo.os(d, "namedLocation")
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
