% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhpOffsets.doc.R
\docType{data}
\name{dhpOffsets}
\alias{dhpOffsets}
\title{Documentation for dhpOffsets data set}
\format{
A data frame with 12 rows and 3 variables
\describe{
  \item{pointID}{identifier of point within plot}
  \item{eastOff}{offset to be added to easting value of southwest corner of plot}
  \item{northOff}{offset to be added to northing value of southwest corner of plot}
  }
}
\usage{
dhpOffsets
}
\description{
Location offsets within NEON plots for locations used to take digital hemispherical photos.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Courtney Meier \email{cmeier@battelleecology.org}
}
\keyword{datasets}
