% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDateMatch.R
\name{findDateMatch}
\alias{findDateMatch}
\title{Find the location data matching the sampling date}
\usage{
findDateMatch(data, locCol, recDate)
}
\arguments{
\item{data}{A data frame including named locations and sampling dates, with duplicated records for locations with a history. This function identifies the record to keep.}

\item{locCol}{The column name of the column containing the named locations}

\item{recDate}{The field name of the date field to be used for matching}
}
\value{
The original data frame with the correct location data added
}
\description{
For a NEON location with a history, find the location matching the sampling date for a given sampling event. Used internally in TOS location calculations.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
