% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocBRD.R
\name{getLocBRD}
\alias{getLocBRD}
\title{Calculate more precise geolocations for bird point counts (DP1.10003.001)}
\usage{
getLocBRD(data, dataProd, token = NA_character_)
}
\arguments{
\item{data}{A data frame containing NEON named locations and other sampling information.}

\item{dataProd}{The table name of the NEON data product table to find locations for.}

\item{token}{User specific API token (generated within neon.datascience user accounts). Optional.}
}
\value{
A data frame of geolocations for the input product and data
}
\description{
Calculation Function. Refine the geolocation data associated with NEON bird observations.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
