% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocBySite.R
\name{getLocBySite}
\alias{getLocBySite}
\title{Download all location data for a given NEON site}
\usage{
getLocBySite(site, type = "site", history = F)
}
\arguments{
\item{site}{The four-letter code of a single NEON site, e.g. 'CLBJ'.}

\item{type}{One of 'site', 'TIS', 'TOS', 'AQU', or 'all', indicating terrestrial instrument locations (TIS), terrestrial observational locations (TOS), aquatic locations (AQU), site-level only, or all available locations. Defaults to site.}

\item{history}{Should the location history be included in the query? T or F, defaults to F.}
}
\value{
A data frame of location data.
}
\description{
Query the API for location data by site. Return all available locations; options for subset to IS or OS locations.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
