% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocCDW.R
\name{getLocCDW}
\alias{getLocCDW}
\title{Calculate more precise geolocations for coarse downed wood tally (DP1.10010.001)}
\usage{
getLocCDW(data, token = NA_character_)
}
\arguments{
\item{data}{A data frame containing NEON named locations and other sampling information.}

\item{token}{User specific API token (generated within neon.datascience user accounts). Optional.}
}
\value{
A data frame of geolocations for the input product and data
}
\description{
Calculation Function. Refine the geolocation data associated with NEON coarse downed wood data.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
