% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocCDWdensity.R
\name{getLocCDWdensity}
\alias{getLocCDWdensity}
\title{Calculate more precise geolocations for coarse downed wood bulk density (DP1.10014.001)}
\usage{
getLocCDWdensity(data, token = NA_character_)
}
\arguments{
\item{data}{A data frame containing NEON named locations and other sampling information.}

\item{token}{User specific API token (generated within data.neonscience.org user accounts). Optional.}
}
\value{
A data frame of geolocations for the input product and data
}
\description{
Calculation Function. Refine the geolocation data associated with NEON coarse downed wood bulk density data.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
