% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocProperties.R
\name{getLocProperties}
\alias{getLocProperties}
\title{Extract location property values from the list extracted from the JSON returned by the NEON API}
\usage{
getLocProperties(locList, history = FALSE)
}
\arguments{
\item{locList}{A list derived from the JSON object returned by the locations endpoint of the NEON API.}

\item{history}{Does locJSON include the location history? T or F, defaults to F.}
}
\value{
A data frame of location data.
}
\description{
Extract location properties from JSON returned by the NEON API. Called within getLocValues(), not exported for independent use.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{internal}
