% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocTOS.R
\name{getLocTOS}
\alias{getLocTOS}
\title{Calculate more precise geolocations for specific NEON data products}
\usage{
getLocTOS(data, dataProd, convertBLAN = FALSE, token = NA_character_)
}
\arguments{
\item{data}{A data frame containing NEON named locations and other sampling information. For reliable results, use data tables as downloaded from the NEON data portal or API.}

\item{dataProd}{The table name of the NEON data product table to find locations for. Refer to package readme for list of possible input data tables.}

\item{convertBLAN}{T or F: Convert locations at BLAN and LEWI from 18N to 17N to match remote sensing data? Defaults to FALSE.}

\item{token}{User specific API token (generated within data.neonscience.org user accounts). Optional.}
}
\value{
A data frame of geolocations for the input product and data
}
\description{
Calculation Function. Refine the geolocation data associated with NEON data products, based on product-specific rules and spatial designs.
}
\details{
In some cases, locations may move slightly over time. The database retains the history, so 
 the location data for a particular named location may include multiple records, one for each time 
 period in the history. When this is the case, getLocTOS() returns the location data matching the 
 collection date in the input data for each record. This means if the same named location is sampled 
 repeatedly over time in the input data, it is possible the location data may differ slightly in 
 different records at the same named location.
}
\examples{
d <- data.frame(uid="e4b111d5", namedLocation="GUAN_044.basePlot.ltr", 
                subplotID=23, trapID="GUAN_044_385")
getLocTOS(d, "ltr_pertrap")
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
