% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocValues.R
\name{getLocValues}
\alias{getLocValues}
\title{Extract location values from the JSON returned by the NEON API}
\usage{
getLocValues(locJSON, history = FALSE)
}
\arguments{
\item{locJSON}{A JSON object returned by the locations endpoint of the NEON API.}

\item{history}{Does locJSON include the location history? T or F, defaults to F.}
}
\value{
A data frame of location data.
}
\description{
Extract location data and properties from JSON returned by the NEON API; designed to be used iteratively for many locations. Called within getLocBySite(), not exported for independent use.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
