% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tLocProps.R
\name{tLocProps}
\alias{tLocProps}
\title{Transpose dataframe of value names and values into single-row dataframe of values with names}
\usage{
tLocProps(locProps)
}
\arguments{
\item{locProps}{A data frame of location properties values as returned by the NEON API.}
}
\value{
A data frame of location properties with columns named by the value names.
}
\description{
Transpose location properties format. Helper function for getLocValues().
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
