% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocByName.R
\name{getLocByName}
\alias{getLocByName}
\title{Extract geolocation data from NEON API}
\usage{
getLocByName(
  data,
  locCol = "namedLocation",
  locOnly = F,
  token = NA_character_
)
}
\arguments{
\item{data}{A data frame in which one column contains the named locations}

\item{locCol}{The column name of the column containing the named locations. Defaults to namedLocation}

\item{locOnly}{Boolean whether to return the full input data frame or just the extracted geolocations}

\item{token}{User specific API token (generated within neon.datascience user accounts). Optional.}
}
\value{
A data frame of the geolocation data for the input named locations
}
\description{
Get geolocation data from the NEON API, for a given set of named location values
}
\examples{
d <- data.frame(namedLocation=c("GUAN_044.basePlot.ltr","GRSM_003.birdGrid.brd"), otherData=c(1,2))
getLocByName(d, "namedLocation")
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
