% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocBySite.R
\name{getLocBySite}
\alias{getLocBySite}
\title{Download all location data for a given NEON site}
\usage{
getLocBySite(site, type = "IS")
}
\arguments{
\item{site}{The four-letter code of a single NEON site, e.g. 'CLBJ'.}

\item{type}{One of 'site', 'IS', 'OS', or 'all', indicating instrumented (IS), observational (OS), site only, or all locations. Defaults to IS.}
}
\value{
A data frame of location data.
}
\description{
Query the API for location data by site. Return all available locations; options for subset to IS or OS locations.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
