% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocTOS.R
\name{getLocTOS}
\alias{getLocTOS}
\title{Calculate more precise geolocations for specific NEON data products}
\usage{
getLocTOS(data, dataProd)
}
\arguments{
\item{data}{A data frame containing NEON named locations and other sampling information. For reliable results, use data tables as downloaded from the NEON data portal or API.}

\item{dataProd}{The table name of the NEON data product table to find locations for. Must be one of: ltr_pertrap, hbp_perbout, sls_soilCoreCollection, brd_perpoint or brd_countdata, mam_pertrapnight, div_1m2Data or div_10m2Data100m2Data, vst_mappingandtagging.}
}
\value{
A data frame of geolocations for the input product and data
}
\description{
Calculation Function. Refine the geolocation data associated with NEON data products, based on product-specific rules and spatial designs.
}
\examples{
d <- data.frame(namedLocation="GUAN_044.basePlot.ltr", subplotID=23, trapID="GUAN_044_385")
getLocTOS(d, "ltr_pertrap")
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
