% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocTOS.R
\name{getLocTOS}
\alias{getLocTOS}
\title{Calculate more precise geolocations for specific NEON data products}
\usage{
getLocTOS(data, dataProd, convertBLAN = FALSE, token = NA_character_)
}
\arguments{
\item{data}{A data frame containing NEON named locations and other sampling information. For reliable results, use data tables as downloaded from the NEON data portal or API.}

\item{dataProd}{The table name of the NEON data product table to find locations for. Refer to package readme for list of possible input data tables.}

\item{convertBLAN}{T or F: Convert locations at BLAN and LEWI from 18N to 17N to match remote sensing data? Defaults to FALSE.}

\item{token}{User specific API token (generated within data.neonscience.org user accounts). Optional.}
}
\value{
A data frame of geolocations for the input product and data
}
\description{
Calculation Function. Refine the geolocation data associated with NEON data products, based on product-specific rules and spatial designs.
}
\examples{
d <- data.frame(namedLocation="GUAN_044.basePlot.ltr", subplotID=23, trapID="GUAN_044_385")
getLocTOS(d, "ltr_pertrap")
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
